Reference

The following reproduces the in-lined documentation about types and methods of the QuickHeaps package. This documentation is also available from the REPL by typing ? followed by the name of a method or a type.

Binary Heaps

QuickHeaps.AbstractBinaryHeap — Type

QuickHeaps.AbstractBinaryHeap{T,O}

is the super-type of binary heaps in QuickHeaps whose values have type T and whose ordering has type O.

The following methods are available for a binary heap h (those which modify the heap contents re-order heap values as needed to maintain the heap structure):

pop!(h)        # deletes and returns root value of heap h
push!(h, x)    # pushes value x in heap h
empty!(h)      # empties heap h
isempty(h)     # yields whether heap h is empty
delete!(h, i)  # deletes i-th value from heap h
peek(h)        # yields root value of heap h without deleting it
first(h)       # idem
setroot!(h, x) # same as h[1] = x, replaces root value of heap h by x

A binary heap h behaves like an abstract vector (with 1-based linear indices), in particular:

length(h)   # the number of values in heap h
h[i]        # the i-th value of heap h
h[i] = x    # set the i-th value of heap h and heapify h

Note that h[1] is the root value of the heap h and that setting a value in the heap may trigger reordering of the values to maintain the binary heap structure. In other words, after doing h[i] = x, do not assume that h[i] yields x.

Operations that modify the heap, like deletion by delete!(h,i), insertion by h[i] = x, pushing by push!(h,x), and extracting by pop!(h) are of complexity O(1) in the best case, O(log(n)) in the worst case, with n = length(h) the number of values in the heap h. Retrieving a given value with peek(h), first(h), or h[i] is always of complexity O(1).

QuickHeaps.BinaryHeap — Type

h = BinaryHeap{T}([o::Base.Order.Ordering = TotalMin,][ vals::AbstractVector])

Build an empty binary heap whose values have type T and with ordering specified by o.

The method QuickHeaps.lt(o,x::T,y::T) is called to determine the order of values x and y in the heap. The default ordering, TotalMin, yields a min-heap object; with TotalMax ordering, a max-heap object is returned.

An optional vector vals storing the initial values of the binary heap can be specified. These values in vals need not be ordered, the BinaryHeap constructor automatically takes care of that. If vals is a Vector{T} instance, the binary-heap will be directly built into vals. Call BinaryHeap(copy(vals)) to create a binary heap with its own storage.

Arguments o and vals may be specified in any order.

Method sizehint!(h,n) may be called to anticipate that the heap may contains n values.

QuickHeaps.FastBinaryHeap — Type

h = FastBinaryHeap{T}([o::Base.Order.Ordering = TotalMin,][ vals::AbstractVector])

Build a fast binary heap. Compared to BinaryHeap{T}(...), the array backing the storage of the heap values is never automatically reduced to improve performances in some cases. You may call resize!(h) to explicitly reduce the storage of fast binary-heap h to its minimum.

QuickHeaps.heapify — Function

heapify([o=TotalMin,] A, n=length(A))

yields an array with the n first values of array A stored in a binary heap structure of ordering specified by o. The storage of the returned heap is a different array than A. Arguments may be specified in any order.

QuickHeaps.heapify! — Function

heapify!(h) -> h

reorders the values in the binary heap h in-place. This method should be called to initialize the heap or to re-order the heap if its contents have been modified by other methods than pop! or push!.

The method can be called at a lower level to heapify (part of) an array storing the heap values:

heapify!([o=TotalMin,] A, n=length(A)) -> A

reorders the n first elements of array A in-place to form a binary heap according to the ordering specified by o. The array A must have 1-based linear indexing. Arguments may be specified in any order.

QuickHeaps.heapify_down! — Function

QuickHeaps.heapify_down!(o, A, i, x=A[i], n=lengh(A)) -> A

stores the value x in the i-th entry of the binary heap built into the n first elements of array A with ordering o and, if needed, moves down the inserted value to maintain the binary heap structure.

This method is called to heapify an array in order to initialize or rebuild the heap structure or to replace the value of the root value of the heap and update the heap structure.

QuickHeaps.heapify_up! — Function

QuickHeaps.heapify_up!(o, A, i, x=A[i]) -> A

stores the value x in the i-th entry of the binary heap built into the i first elements of array A with ordering o and, if needed, moves up the value to maintain the heap structure.

QuickHeaps.isheap — Function

isheap([o=TotalMin,], A, n=length(A))

yields whether the n first elements of array A have a binary heap structure ordered as specified by o. Arguments may be specified in any order.

isheap(obj; check=false)

yields whether object obj is a binary heap. If keyword check is true, the internal structure of obj is checked; otherwise, the type of obj is trusted to determine whether it is a binary heap.

QuickHeaps.unsafe_heapify_down! — Function

QuickHeaps.unsafe_heapify_down!(o, A, i, x=A[i], n=lengh(A))

This method is a fast but unsafe version of QuickHeaps.heapify_down! which assumes that all arguments are correct, that is A implements 1-based linear indexing, 0 ≤ n ≤ lengh(A), and 1 ≤ i ≤ n.

QuickHeaps.unsafe_heapify_up! — Function

QuickHeaps.unsafe_heapify_up!(o, A, i, x=A[i])

This methods is a fast but unsafe version of QuickHeaps.heapify_up! which assumes that all arguments are correct, that is A implements 1-based linear indexing and 1 ≤ i ≤ length(A).

QuickHeaps.unsafe_grow! — Function

QuickHeaps.unsafe_grow!(h, n) -> A

grows the size of the binary heap h to be n and returns the array A backing the storage of the values. This method is unsafe because it does not check its arguments and because it breaks the binary heap structure of the array of values.

This method is called by push! to grow the size of the heap and shall be specialized for any concrete sub-types of QuickHeaps.AbstractBinaryHeap.

QuickHeaps.unsafe_grow!(pq, n) -> pq

grows the size of the binary heap backing the storage of the nodes of the priority queue pq to be n and returns the priority queue object.

QuickHeaps.unsafe_shrink! — Function

QuickHeaps.unsafe_shrink!(h, n)

shrinks the size of the binary heap h to be n. This method is unsafe because it does not check its arguments.

This method is called by delete! to eventually reduce the size of the heap and shall be specialized for any concrete sub-type of QuickHeaps.AbstractBinaryHeap.

QuickHeaps.unsafe_shrink!(pq, n)

shrinks the size of the binary heap backing the storage of the nodes of the priority queue pq to be n.

Priority Queues

QuickHeaps.AbstractPriorityQueue — Type

QuickHeaps.AbstractPriorityQueue{K,V,O}

is the super type of priority queues with ordering of type O<:Base.Order.Ordering and storing nodes associating a key of type K with a priority value of type V.

Package QuickHeaps provides two concrete types of priority queues: PriorityQueue for any kind of keys and FastPriorityQueue for which keys are analogous to array indices.

Priority queues behave like dictionaries with the additional feature of automatically maintaining an ordered structure according to the priority queue ordering and the node values. For a priority queue pq, retrieving the root node (i.e., the one of highest priority) without removing it costs O(1) and is done by:

peek(pq, T=Pair) -> T(key, val)

with T the type of the expected result. Retrieving the value of a node given its key has also an O(1) complexity and is done by one of:

pq[key...] -> val
getindex(pq, key...) -> val
get(pq, key, def) -> val_at_key_or_def

Changing the content of the priority queue has a complexity of O(log(n)) with n = length(pq) the number of nodes in the queue. This includes removing the node at key by:

delete!(pq, key) -> pq

or removing the root node by:

pop!(pq)          # -> root node as a `key=>val` pair
dequeue!(pq)      # -> key of root node
dequeue_pair!(pq) # -> root node as a `key=>val` pair
dequeue_node!(pq) # -> root node as stored in priority queue

or setting a node x with a given key and value val by one of:

pq[key] = val
enqueue!(pq, key, val)
enqueue!(pq, key => val)
push!(pq, key => val)
push!(pq, (key, val))
push!(pq, x) # x is a node with a key and a value

QuickHeaps.PriorityQueue — Type

PriorityQueue{K,V}([o=TotalMin,] T=Node{K,V})

yields a priority queue for keys of type K and priority values of type V. Optional arguments o::Ordering and T<:AbstractNode{K,V} are to specify the ordering of values and type of nodes to store key-value pairs. Type parameters K and V may be omitted if the node type T is specified.

Having a specific node type may be useful to specialize the QuickHeaps.lt method which is called to determine the order.

If keys are analogous to array indices (linear or Cartesian), FastPriorityQueue may provide a faster alternative.

QuickHeaps.FastPriorityQueue — Type

FastPriorityQueue{V}([o=TotalMin,] [T=Node{Int,V},] dims...)

yields a priority queue for keys analogous of indices in an array of size dims... and priority values of type V. Optional arguments o::Ordering and T<:AbstractNode{Int,V} are to specify the ordering of values and type of nodes to store key-value pairs (the key is stored as a linear index of type Int). Type parameter V may be omitted if the node type T is specified.

See PriorityQueue if keys cannot be assumed to be array indices.

DataStructures.dequeue! — Method

dequeue!(pq) -> key

Remove the root node from the priority queue pq and return its key.

You may callnode!(pq)`](@ref dequeuenode!) or dequeue_pair!(pq) to dequeue the root node as stored in pq or as a key-value pair.

QuickHeaps.dequeue_node! — Function

dequeue_node!(pq) -> node

Removes and return the root node from the priority queue pq.

Also see dequeue! and dequeue_pair!.

DataStructures.dequeue_pair! — Function

dequeue_pair!(pq) -> (key => val)

removes the root node from the priority queue pq and returns it as a key-value Pair. This is the same as pop!(pq).

Also see dequeue! and dequeue_node!.

DataStructures.enqueue! — Method

pq[key...] = val
setindex!(pq, val, key...) -> pq
enqueue!(pq, key, val) -> pq
enqueue!(pq, key=>val) -> pq
push!(pq, key=>val) -> pq

Set the value val stored by the priority queue pq at index key automatically maintaining the partial ordering of pq.

If x is an object which has a key and a value retrievable by QuickHeaps.get_key(x) and QuickHeaps.get_val(x), then the following calls are other possibilities:

enqueue!(pq, x) -> pq
push!(pq, x) -> pq

Nodes

QuickHeaps.AbstractNode — Type

QuickHeaps.AbstractNode{K,V}

is the super-type of nodes with a key of type K and a value of type V. Nodes can be used in binary heaps and priority queues to represent key-value pairs and specific ordering rules may be imposed by specializing the QuickHeaps.lt function which, for abstract nodes, is by default:

QuickHeaps.lt(o::Base.Order.Ordering, x::T, y::T) where {T<:QuickHeaps.AbstractNode} =
    Base.Order.lt(o, QuickHeaps.get_val(x), QuickHeaps.get_val(y))

QuickHeaps.Node — Type

QuickHeaps.Node{K=typeof(k),V=typeof(v)}(k,v)

yields a node storing key k and value v. Optional type parameters K and V are the respective types of the key and of the value.

See also QuickHeaps.AbstractNode, QuickHeaps.AbstractPriorityQueue.

QuickHeaps.get_key — Function

QuickHeaps.get_key(x::QuickHeaps.AbstractNode) -> k

yields the key k of node x. This method may be specialized for any sub-types of QuickHeaps.AbstractNode.

Also see QuickHeaps.get_val.

QuickHeaps.get_val — Function

QuickHeaps.get_val(x::QuickHeaps.AbstractNode) -> v

yields the value v of node x. This method may be specialized for any sub-types of QuickHeaps.AbstractNode.

Also see QuickHeaps.get_key.

Orderings

QuickHeaps.FastMin — Constant

QuickHeaps.FastMinOrdering <: Base.Order.Ordering
const FastMin = QuickHeaps.FastMinOrdering()

Singleton for min ordering. This ordering is faster than TotalMin but leaves indefinite the order of NaN values.

QuickHeaps.FastMinOrdering is the type of FastMin.

Also see FastMax.

QuickHeaps.FastMax — Constant

FastMax

Singleton for max ordering. This ordering is faster than TotalMax but leaves indefinite the order of NaN values.

QuickHeaps.FastMaxOrdering is an alias to the type of FastMax.

Also see FastMin.

QuickHeaps.TotalMin — Constant

TotalMin

Singleton for total min ordering considering NaN's as greater than any other floating-point value, and missing to be greater than anything else. With this ordering, values are sorted in ascending order, followed by NaN then missing values.

TotalMin is similar to the default ordering in most Julia algorithms and which is implemented by isless. However, for arrays of floating-point values with 50% of NaN's, TotalMin is nearly as fast as (19%) using < and is almost twice faster (82%) than isless. This speed-up is obtained by using non-branching bitwise operators instead of logical operators.

QuickHeaps.TotalMinOrdering is the type of TotalMin.

Also see TotalMax.

QuickHeaps.TotalMax — Constant

TotalMax

Singleton for total max ordering. With this ordering, values are sorted in decreasing order, followed by NaN then missing values.

QuickHeaps.TotalMaxOrdering is the type of TotalMax.

Note

TotalMax and reverse(TotalMin) both sort regular values in decreasing order but the latter puts missing then NaN values first.

Also see TotalMin.

Miscellaneous

The following non-exported methods may be needed for implementing new types of binary heap or of priority queue. End-users probably not have to worry about these.

QuickHeaps.has_bad_values — Function

QuickHeaps.has_bad_values(A[, isbad])

yields whether array A has bad values according to predicate isbad. For arrays with floating-point values, isbad default to isnan if unspecified. For integer-valued arrays, this function always returns false if isnan is unspecified.

QuickHeaps.has_standard_linear_indexing — Function

QuickHeaps.has_standard_linear_indexing(A)

yields whether array A implements standard linear indexing (1-based).

QuickHeaps.heap_index — Function

QuickHeaps.heap_index(pq, key) -> i::Int

Return the index i corresponding to key in the binary heap backing the storage of the nodes of the priority queue pq. If the key is not in priority queue, i = 0 is returned, otherwise i ∈ 1:n with n = length(pq) is returned.

The heap_index method is used to implement haskey, get, and delete! methods for priority queues. The heap_index method shall be specialized for any concrete sub-types of QuickHeaps.AbstractPriorityQueue.

QuickHeaps.in_range — Function

QuickHeaps.in_range(i, len::Integer)

yields whether 1 ≤ i ≤ len.

QuickHeaps.in_range(i, A::Array)

yields whether i is a valid linear index of array A.

QuickHeaps.in_range(i, R::AbstractUnitRange{<:Integer})

yields whether i is in the range R.

QuickHeaps.is_one_based_unit_range — Function

QuickHeaps.is_one_based_unit_range(itr)

yields whether iterator itr is a 1-based unit range.

QuickHeaps.linear_index — Function

QuickHeaps.linear_index(pq, key) -> k::Int

Convert key into a linear index suitable for the fast priority queue pq. The key can be a linear index or a multi-dimensional index (anything accepted by to_indices). The current settings for bounds checking are used.

QuickHeaps.lt — Function

QuickHeaps.lt(o::Ordering, x::T, y::T)

Return whether x is less than y according to ordering o.

QuickHeaps.lt is called by QuickHeaps to build ordered structures like binary heaps and priority queues. For T <: QuickHeaps.AbstractNode, QuickHeaps.lt compares the values of the nodes x and y; otherwise, QuickHeaps.lt calls Base.Order.lt to compare values.

QuickHeaps.nodes — Function

QuickHeaps.nodes(pq)

Yield the vector backing the storage of the nodes of priority queue pq. This vector has a binary-heap structure.

QuickHeaps.index — Function

QuickHeaps.index(pq) -> I

Yield the key to index mapping of priority queue pq such that I[key] is the index of the node associated with key in the binary heap QuickHeaps.nodes(pq).

QuickHeaps.storage — Function

QuickHeaps.storage(h)

yields the array backing the storage of the values in the binary heap h.

This method may be specialized for custom binary heap types.

QuickHeaps.ordering — Function

QuickHeaps.ordering(A) -> o

Return the object o specifying the ordering of the values in the object A, a binary heap or a priority queue.

This method may be specialized for custom types.

QuickHeaps.setroot! — Function

setroot!(h, x) -> h

replaces the value of the root note in heap h by x. This is similar to h[1] = x but a bit faster.

QuickHeaps.to_key — Function

QuickHeaps.to_key(pq, key)

Convert key to the type of keys used by the priority queue pq.

QuickHeaps.to_node — Function

QuickHeaps.to_node(pq, key => val)
QuickHeaps.to_node(pq, key, val)

Convert the association of key and value val into a node of the type stored by the priority queue pq.

QuickHeaps.to_node(pq, x)

Convert x into a node of the type stored by the priority queue pq. Methods QuickHeaps.get_key(x) and QuickHeaps.get_val(x) must be implemented for object x,

QuickHeaps.to_val — Function

QuickHeaps.to_val(pq, val)

Convert value val to the type of values stored by the priority queue pq.

QuickHeaps.typename — Function

QuickHeaps.typename(x)
QuickHeaps.typename(typeof(x))

yield a short string describing the type of object x.